+++ layers/CMSSchema/packages/categories/src/TypeAPIs/QueryableTaxonomyCategoryListTypeAPIInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PoPCMSSchema\Categories\TypeAPIs;
-
-interface QueryableTaxonomyCategoryListTypeAPIInterface extends TaxonomyCategoryListTypeAPIInterface
-{
-}