+++ layers/CMSSchema/packages/tags/src/TypeAPIs/InjectableTaxonomyTagListTypeAPIInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPCMSSchema\Tags\TypeAPIs;
+
+interface InjectableTaxonomyTagListTypeAPIInterface extends TaxonomyTagListTypeAPIInterface
+{
+}